Associações das Models

+++ src/database/models/Adress.js
@@ -27,7 +27,7 @@ module.exports = (sequelize, DataTypes) => {
             type: DataTypes.STRING(100),
         },
         street_number:{
-            type: DataTypes.FLOAT,
+            type: DataTypes.DECIMAL(6,2),
             allowNull: false
         },
         is_current:{
